Extract `InMemoryAppender`
This commit extracts the `InMemoryAppender` into the test utils package. This is a 1st step that would be followed by extracting this class completely to an independent tests utilities library that would be shared across the project. (see Issue #55 for more details)
